HV: hypercall: make hypercall functions return int32_t
The error code in the hypervisor is 32-bit signed integers. To reduce implicit
conversions, this patch make hcall_xxx returns int32_t, and finally converts it
to uint64_t when assigned to rax whose semantics is properly defined in C99.

Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/arch/x86/guest/vmcall.c
@@ -14,7 +14,7 @@
  */
 int vmcall_vmexit_handler(struct vcpu *vcpu)
 {
-	int64_t ret = -EACCES;
+	int32_t ret = -EACCES;
 	struct vm *vm = vcpu->vm;
 	struct run_context *cur_context =
 		&vcpu->arch_vcpu.contexts[vcpu->arch_vcpu.cur_context];
