+++ hypervisor/include/common/hypercall.h
@@ -37,7 +37,7 @@ int acrn_vpic_inject_irq(struct vm *vm, int irq, enum irq_mode mode);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_get_api_version(struct vm *vm, uint64_t param);
+int32_t hcall_get_api_version(struct vm *vm, uint64_t param);
 
 /**
  * @brief create virtual machine
@@ -52,7 +52,7 @@ int64_t hcall_get_api_version(struct vm *vm, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_create_vm(struct vm *vm, uint64_t param);
+int32_t hcall_create_vm(struct vm *vm, uint64_t param);
 
 /**
  * @brief destroy virtual machine
@@ -64,7 +64,7 @@ int64_t hcall_create_vm(struct vm *vm, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_destroy_vm(uint64_t vmid);
+int32_t hcall_destroy_vm(uint64_t vmid);
 
 /**
  * @brief resume virtual machine
@@ -77,7 +77,7 @@ int64_t hcall_destroy_vm(uint64_t vmid);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_resume_vm(uint64_t vmid);
+int32_t hcall_resume_vm(uint64_t vmid);
 
 /**
  * @brief pause virtual machine
@@ -90,7 +90,7 @@ int64_t hcall_resume_vm(uint64_t vmid);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_pause_vm(uint64_t vmid);
+int32_t hcall_pause_vm(uint64_t vmid);
 
 /**
  * @brief create vcpu
@@ -106,7 +106,7 @@ int64_t hcall_pause_vm(uint64_t vmid);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_create_vcpu(struct vm *vm, uint64_t vmid, uint64_t param);
+int32_t hcall_create_vcpu(struct vm *vm, uint64_t vmid, uint64_t param);
 
 /**
  * @brief assert IRQ line
@@ -121,7 +121,7 @@ int64_t hcall_create_vcpu(struct vm *vm, uint64_t vmid, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_assert_irqline(struct vm *vm, uint64_t vmid, uint64_t param);
+int32_t hcall_assert_irqline(struct vm *vm, uint64_t vmid, uint64_t param);
 
 /**
  * @brief deassert IRQ line
@@ -136,7 +136,7 @@ int64_t hcall_assert_irqline(struct vm *vm, uint64_t vmid, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_deassert_irqline(struct vm *vm, uint64_t vmid, uint64_t param);
+int32_t hcall_deassert_irqline(struct vm *vm, uint64_t vmid, uint64_t param);
 
 /**
  * @brief trigger a pulse on IRQ line
@@ -151,7 +151,7 @@ int64_t hcall_deassert_irqline(struct vm *vm, uint64_t vmid, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_pulse_irqline(struct vm *vm, uint64_t vmid, uint64_t param);
+int32_t hcall_pulse_irqline(struct vm *vm, uint64_t vmid, uint64_t param);
 
 /**
  * @brief inject MSI interrupt
@@ -165,7 +165,7 @@ int64_t hcall_pulse_irqline(struct vm *vm, uint64_t vmid, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_inject_msi(struct vm *vm, uint64_t vmid, uint64_t param);
+int32_t hcall_inject_msi(struct vm *vm, uint64_t vmid, uint64_t param);
 
 /**
  * @brief set ioreq shared buffer
@@ -180,7 +180,7 @@ int64_t hcall_inject_msi(struct vm *vm, uint64_t vmid, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_set_ioreq_buffer(struct vm *vm, uint64_t vmid, uint64_t param);
+int32_t hcall_set_ioreq_buffer(struct vm *vm, uint64_t vmid, uint64_t param);
 
 /**
  * @brief notify request done
@@ -193,7 +193,7 @@ int64_t hcall_set_ioreq_buffer(struct vm *vm, uint64_t vmid, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_notify_req_finish(uint64_t vmid, uint64_t param);
+int32_t hcall_notify_req_finish(uint64_t vmid, uint64_t param);
 
 /**
  * @brief setup ept memory mapping
@@ -208,7 +208,7 @@ int64_t hcall_notify_req_finish(uint64_t vmid, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_set_vm_memmap(struct vm *vm, uint64_t vmid, uint64_t param);
+int32_t hcall_set_vm_memmap(struct vm *vm, uint64_t vmid, uint64_t param);
 
 /**
  * @brief setup ept memory mapping for multi regions
@@ -219,7 +219,7 @@ int64_t hcall_set_vm_memmap(struct vm *vm, uint64_t vmid, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_set_vm_memmaps(struct vm *vm, uint64_t param);
+int32_t hcall_set_vm_memmaps(struct vm *vm, uint64_t param);
 
 /**
  * @brief remap PCI MSI interrupt
@@ -234,7 +234,7 @@ int64_t hcall_set_vm_memmaps(struct vm *vm, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_remap_pci_msix(struct vm *vm, uint64_t vmid, uint64_t param);
+int32_t hcall_remap_pci_msix(struct vm *vm, uint64_t vmid, uint64_t param);
 
 /**
  * @brief translate guest physical address to host physical address
@@ -248,7 +248,7 @@ int64_t hcall_remap_pci_msix(struct vm *vm, uint64_t vmid, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_gpa_to_hpa(struct vm *vm, uint64_t vmid, uint64_t param);
+int32_t hcall_gpa_to_hpa(struct vm *vm, uint64_t vmid, uint64_t param);
 
 /**
  * @brief Assign one passthrough dev to VM.
@@ -260,7 +260,7 @@ int64_t hcall_gpa_to_hpa(struct vm *vm, uint64_t vmid, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_assign_ptdev(struct vm *vm, uint64_t vmid, uint64_t param);
+int32_t hcall_assign_ptdev(struct vm *vm, uint64_t vmid, uint64_t param);
 
 /**
  * @brief Deassign one passthrough dev from VM.
@@ -272,7 +272,7 @@ int64_t hcall_assign_ptdev(struct vm *vm, uint64_t vmid, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_deassign_ptdev(struct vm *vm, uint64_t vmid, uint64_t param);
+int32_t hcall_deassign_ptdev(struct vm *vm, uint64_t vmid, uint64_t param);
 
 /**
  * @brief Set interrupt mapping info of ptdev.
@@ -284,7 +284,7 @@ int64_t hcall_deassign_ptdev(struct vm *vm, uint64_t vmid, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_set_ptdev_intr_info(struct vm *vm, uint64_t vmid, uint64_t param);
+int32_t hcall_set_ptdev_intr_info(struct vm *vm, uint64_t vmid, uint64_t param);
 
 /**
  * @brief Clear interrupt mapping info of ptdev.
@@ -296,7 +296,7 @@ int64_t hcall_set_ptdev_intr_info(struct vm *vm, uint64_t vmid, uint64_t param);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_reset_ptdev_intr_info(struct vm *vm, uint64_t vmid,
+int32_t hcall_reset_ptdev_intr_info(struct vm *vm, uint64_t vmid,
 	uint64_t param);
 
 /**
@@ -308,7 +308,7 @@ int64_t hcall_reset_ptdev_intr_info(struct vm *vm, uint64_t vmid,
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_setup_sbuf(struct vm *vm, uint64_t param);
+int32_t hcall_setup_sbuf(struct vm *vm, uint64_t param);
 
 /**
  * @brief Get VCPU Power state.
@@ -320,7 +320,7 @@ int64_t hcall_setup_sbuf(struct vm *vm, uint64_t param);
  * @return 0 on success, non-zero on error.
  */
 
-int64_t hcall_get_cpu_pm_state(struct vm *vm, uint64_t cmd, uint64_t param);
+int32_t hcall_get_cpu_pm_state(struct vm *vm, uint64_t cmd, uint64_t param);
 
 /**
  * @defgroup trusty_hypercall Trusty Hypercalls
@@ -346,7 +346,7 @@ int64_t hcall_get_cpu_pm_state(struct vm *vm, uint64_t cmd, uint64_t param);
  * @return 0 on success, non-zero on error.
  */
 
-int64_t hcall_world_switch(struct vcpu *vcpu);
+int32_t hcall_world_switch(struct vcpu *vcpu);
 
 /**
  * @brief Initialize environment for Trusty-OS on a vCPU.
@@ -362,7 +362,7 @@ int64_t hcall_world_switch(struct vcpu *vcpu);
  *
  * @return 0 on success, non-zero on error.
  */
-int64_t hcall_initialize_trusty(struct vcpu *vcpu, uint64_t param);
+int32_t hcall_initialize_trusty(struct vcpu *vcpu, uint64_t param);
 
 /**
  * @}
